Refactor controller functions into separate file
