+++ controllers/controllers.js
@@ -1,0 +1,4 @@
+module.exports = {
+  getSightings: require('./getSightings'),
+  getAnimals: require('./getAnimals')
+};
